Started big overhaul to simplify the logic, see commit message
so far, test_transition_sequences is passing

Major change
============

Functions in logic.py no longer return lists of Actions that can be
invoked later, they now call action functions immediately when they're
executed.  The reason for that big change is that code started to be
complex, the payoff for the approach used was minimal, and show-stopper
was that targets of Choice transitions couldn't be evaluated when
they're encountered in the sequence since evaluation might require an
updated value of some variable, and that update isn't performed yet
since actions that update them were NOT being executed immediately.

This change impacts testability a bit, which is unfortunate, but this is
after all a *state* machine, stateless approach can only get you so far.
Most functions used to be pure (without side-effects), and returned
sequences which were easy to check. Now some of the functions are
implemented in a way that rely on side-effects, and testing strategy is
now to use LoggingState and LoggingTransitions provided in
tests/reusable.py that append new entry to hsm.data._log list when they
are called, and have tests check values of that list.

These changes are reflected in renaming of the functions:
  * `get_merged_sequences` -> `perform_transition`
  * `get_response_sequence` -> `perform_response`
  * `entry_sequence` -> `enter`

Other changes
=============

* State entry and exit actions now take only one argument (HSM instance)

* all Transitions are now subclasses of _GenericTransition, makes code
  simpler since all transitions are assumed to have multiple targets
  specified in `switch` dictionary, and target is chosen using
  `get_target` method. Transition and LocalTransition are a special case
  whose `switch` dict has only one item and `get_target` always returns
  that item. Similarly, InternalTransition has an empty `switch` dict
  and `get_target` returns None.

* transitions now also have a `source` property whose value is sig of
  the state from which transition starts. It's only used for testing.

* removed Action, no need for them anymore